AdvancedAgents/ShadowARB/requirements.txt
# Core Dependencies
langgraph==0.2.45
langchain-core==0.3.15
PyGithub==2.4.0
pydantic==2.9.2
litellm>=1.52.0
python-dotenv==1.0.1
tiktoken>=0.7.0
